feat: User.expCur/expNext를 long으로 변경 (DB NUMBER(12) 대응, 오버플로우 방지)
- User.java: expCur, expNext int → long
- BossAttackController: applyExpAndLevelUp/LevelUpResult/calcNextExp long 처리
- BotNewDAO/Service/ServiceImpl: updateUserAfterBattleTx 파라미터 long으로 통일

+++ src/main/java/my/prac/core/game/dto/User.java
@@ -12,8 +12,8 @@ public  class User {
 	public String userName;
 	public String roomName;
 	public int lv;
-	public int expCur;
-	public int expNext;
+	public long expCur;
+	public long expNext;
 	public int hpCur;
 	public int hpMax;
 	public int hpRegen;
